refactor: uploader 경로 변경, S3ClientConfig config 폴더로 이동 (#255)
